+++ ReactVersion/src/types.ts
@@ -1,0 +1,8 @@
+export interface Hell {
+  description: string;
+  observed: Date;
+}
+
+export interface Stats {
+  count: number;
+}